tests: Clean up after view controller refactoring
