+++ phpunit.dist.xml
@@ -27,26 +27,10 @@
 		</include>
 
 		<exclude>
-			<directory suffix=".php">./config/areas/lab</directory>
+			<directory suffix=".php">./config/areas</directory>
 			<directory suffix=".php">./config/blocks</directory>
 			<directory suffix=".php">./config/templates</directory>
 			<file>./config/api/routes/changes.php</file>
-			<file>./config/areas/account/buttons.php</file>
-			<file>./config/areas/account/dropdowns.php</file>
-			<file>./config/areas/files/buttons.php</file>
-			<file>./config/areas/files/requests.php</file>
-			<file>./config/areas/lab/drawers.php</file>
-			<file>./config/areas/languages/buttons.php</file>
-			<file>./config/areas/site/buttons.php</file>
-			<file>./config/areas/site/dropdowns.php</file>
-			<file>./config/areas/site/dialogs.php</file>
-			<file>./config/areas/site/requests.php</file>
-			<file>./config/areas/site/searches.php</file>
-			<file>./config/areas/system/dialogs.php</file>
-			<file>./config/areas/users/dropdowns.php</file>
-			<file>./config/areas/users/buttons.php</file>
-			<file>./config/areas/users/dialogs.php</file>
-			<file>./config/areas/users/searches.php</file>
 			<file>./config/aliases.php</file>
 			<file>./config/setup.php</file>
 		</exclude>
