--- tests/Panel/Areas/LoginTest.php
@@ -1,28 +1,0 @@
-<?php
-
-namespace Kirby\Panel\Areas;
-
-class LoginTest extends AreaTestCase
-{
-	public function setUp(): void
-	{
-		parent::setUp();
-		$this->install();
-	}
-
-	public function testLoginRedirectFromHome(): void
-	{
-		$this->assertRedirect('/', 'login');
-
-		// last path gets remembered
-		$this->assertSame('', $this->app->session()->get('panel.path'));
-	}
-
-	public function testLoginRedirectFromAnywhere(): void
-	{
-		$this->assertRedirect('somewhere', 'login');
-
-		// last path gets remembered
-		$this->assertSame('somewhere', $this->app->session()->get('panel.path'));
-	}
-}
